AI Internz Learn Internship/resume_screening_tool/requirements.txt
flask
nltk
spacy
scikit-learn
bert-extractive-summarizer
pdfminer.six
python-docx
